Added grammar production for CONTAINS statement.
Added flag contains_sect in fortran.y which is set when a CONTAINS
statement was just parsed, then cleared when the following subprogram 
statement is parsed, until exit from the subprogram, when it is 
restored so the context of the next statement will be the same as 
after a CONTAINS statement.

Added enum type SUBPROG_TYPE, and new field subprogtype of this type in
BlockStack struct, in fortran.y.  This field is used by pop_block to
restore the contains_sect flag upon exit from a subprogram within a
CONTAINS section.  It is also used when a CONTAINS statement is parsed,
to check that it is in an allowed context.

Added code to treat MODULE as start of program unit.  This version
does not pass "make check" due to symbol table inconsistencies due
to not yet handling nested scopes: crashes during wrapup.