Change to exprtype.c: corrected propagation of EVALUATED_EXPR flag of result token in binexpr_type() so that the flag is made true only when the value of the binary expression is in fact evaluated.  Formerly it was being made true whenever it was true for both operands. Presently only constant expressions of integer type have their values propagated.
Extended rule clean to remove core.* files in Makefile.in and
test/Makefile.in.  Added make_blockmatch.pl to dependency of
block_match.h in Makefile.in.

Fixed buffer overrun introduced in makedcls.c in revision 4:
Since "prog unit" has 3 more characters than "module", the
wording change required shrinking the space available in the
declaration file headings for the program unit name from 20 to 18
characters and removing one blank between the unit name and the file
name, in order to keep the line length to 72 columns, the size
of the begin array.