Removed or altered C++ syntax to make the code ANSI C compatible.
Changed method of getting type names from direct lookup in type_name
array, which will not work when derived types are involved, to
a function call.  The name array is now elementary_type_name, and
the function is type_name, which returns either the array element
for elementary types or the derived type name for derived types.

First semi-working version of find_Dtype which looks for an existing
derived type definition matching a given one.  Right now it only
looks for the same name.

Added a routine to handle the situation where a TYPE definition is
the first statement of a program, implying a %MAIN program declaration.
In this case it is necessary to reorder the entries in the local
symbol table to match what would have been done if the program
declaration were explicit.  Added two test files to regression
suite, one of which exhibits this case and the other has the
explicit PROGRAM statement: type-first-stmt.f and type-notfirst-stmt.f.

Started adding semantic processing to deal with PRIVATE and SEQUENCE
attributes of derived types.