In dtypes.c: Fixed previous bug which never set the line_declared field and module name (if any) of derived types that first occurred as forward references.  Also fixed bug whereby variables or functions accidentally referenced in a type definition (e.g. misspelled or undeclared parameters in array dimensions or in initializers) would appear as components of the type.
In dtypes.h: added field symt to Dtype struct to point back to local
symbol table entry referring to the type in the current scope.  This
is needed when printing out data type names, to use the local symbol
table entry instead of the name stored in the dtype_table. This
matters when derived types get renamed in RENAME lists.
 
In symtab.c: fixed bug in move_outside_scope that caused it to do the
wrong thing if the moved item was not last in symbol table.

Semantic processing of components with pointer attribute: pointer
attribute is carried in token.

In grammar, unified productions for type declarations of derived type
with those of elementary type.  They formerly had parallel grammars.
Now derived_type_name is one of the options for type_name and the rest
of the declaration is handled with the same productions.

In grammar, added acceptance of pointer assignments in data type
statement initializers.  Thus, both '=' and '=>' are now allowed.
Formerly (in 3.3.1) only '=' was accepted.  With the unified
productions, both forms of initialization can now be present in either
derived or elementary type declarations.

Moved intrinsic function evaluator functions (ii_abs and the like) out
of exprtype.c and into intrins.c.  Changed the IntrinsInfo struct so
each entry holds a pointer to the evaluator if any.  Got rid of the
old ii_fun array that used to provide these pointers.  The old
mechanism was error prone when adding new intrinsics.

Fixed previous bug which failed to set used_before_allocation flag of
ALLOCATABLE types.

Fixed previous bug which failed to accept POINTER or ALLOCATABLE types in
ALLOCATE and DEALLOCATE intrinsic functions.  Variables without DIMENSION
attribute were being skipped due to incorrect logic.

Fixed previous bug which failed to propagate association status in pointer
assignment statements.

In symtab.c : added function do_nullify() which processes NULLIFY
statement.