In dtypes.c : In ref_component, the symbol table entry of base variable is no longer updated with usage information because these actions are done in other places.
In symtab.h : Tokens now have ASSOCIATED_EXPR and ALLOCATED_EXPR
flags.  Changed checking of associated and allocated status of
assigned values to be based on token flags instead of symbol table
entry flags.  New routine set_attr_flags in fortran.y sets the token
flags to match symbol table, to be called when a variable enters the
parse tree.  This change is needed since a non-variable, such as an
invocation of a function returning a pointer, can be assigned.

In exprtype.c : In changed interaction of check_initializer_type and
assignment_stmt_type, to do the right thing for warnings on both the
bastard initializers (combined type and data statements like INTEGER A
/ 10 /) and F90 standard initializers (like INTEGER :: A=10).  The
issue is that in the bastard initializers, the column number for
warnings should be that of the assigned value, since there can be a
list of values assigned, whereas in standard initializers there is
only one value so the column can be that of the = sign.  Formerly
check_initializer_type fudged by passing the assigned value token as the
"equals" argument, but now we need to distinguish '=' from '=>' so it
needs to pass the true equals token.  Now assignment_stmt_type looks
at the equals token, and if it is '/', adjusts the line and column
numbers for reports to be that of term2; otherwise it is that of
equals like in assignment statements.

Fixed previous bug where TARGET variables were not accepted as right
hand side variables in assignment stmts.

Added semantic processing of NULL intrinsic function.  Added
acceptance of derived type arguments for ASSOCIATED and ALLOCATED
intrinsic functions.

In fortran.y : Added grammar for initializers in derived type
declarations.  (However, structure constructors are not yet
supported.)  Non terminal primary now aggregated into two non
terminals : data_object which can occur in ALLOCATE, DEALLOCATE, and
NULLIFY stmts and compound_object which cannot.