In fortran.y : Replaced the non terminal 'derived_type_decl_list' with 'arith_type_decl_list' for derived type declarations.  This unifies the treatment of the declaration lists, avoiding repeated code. Added action to transfer symbol table flags to token for checking allocated/associated status etc.
In symtab.h : Added flag 'assigned_as_target' to local symbol table entry
which gets set when a pointer/target variable is used as the rvalue in
a pointer assignment statement.  This allows suppression of warnings
that said target is not set or used, since we cannot tell if it was
accessed via the pointer instead.  The work is done by the new function
'use_target'.  Added ALLOCATABLE_EXPR to Token struct which identifies
allocatable variables.

In symtab.c : Modified do_allocate, do_deallocate and do_nullify functions
so that they check the token flags instead of referring to the symbol
table entries.  This is necessary for handling components.  Note that
we still are holding allocation and association status in the base
variable and not in individual components.

In dtypes.h : Added allocatable flag to DtypeComponent struct which
identifies allocatable components.

If needed, a key is printed explaining the meaning of ^ < and - in symbol
table listings.