Fixed a bug introduced in a previous revision which caused failure to evaluate some intrinsic functions with constant arguments.
Made some changes to design of how checking of intrinsic function
usage is done, to make it cleaner and easier to maintain and extend.

Added intrins_eval.f90 to the regression test which tests evaluation of
integer intrinsic functions and special handling of some intrinsics.

In Makefile.in : Revised to include datarootdir as the installation
destination for dc2inc.awk.  This reflects modern conventions for
install target.  Changed dependency of block_match.pl on
make_blockmatch.pl to make_blockmatch.pl.in.  The script
make_blockmatch.pl is remade by configure, so the dependency was
causing block_match.pl to be remade after every configure; this could
be a problem for users building ftnchek without a perl installation.
The file make_blockmatch.pl.in is what should be edited when
substantive changes are made.

In symtab.c : Removed leftover function do_bind_spec2.

Fixed incorrect use of type_t to declare things that are only type
(type index).  Use type_t for objects containing type_pack(class,type)
and int for objects containing just the type.  This difference
currently is not significant but it will matter when type_t becomes a
struct.