Fixed do_result_spec so that when type of a function is declared, the result variable now gets this type.  Also for entry point, if its result variable was declared, entry gets that type.
Added routine equivalence_result_vars which equivalences result
variables of a function and its entry statements. If RESULT suffix is
missing, the function/entry variable itself acts as the result
variable and gets equivalenced.

Added is_defn flag to struct TokenListHeader to distinguish a function
or subroutine definition.  This allows definition of a recursive
subprogram to be distinguished from recursive calls.

In prlists.c : In function process_lists, unified code for subprog
definitions and subprog calls, to allow correct processing of
functions with result variables (which now have class_SUBPROGRAM) and
recursive calls.