In symtab.h : Created a new struct 'array_dim_t' which holds array related information such as dimensions, size and size_unknown flag. Replaced 'array_dim' field by this struct in Token struct to hold array dimension information.
The sizes of array-valued expressions are now propagated through
tokens, to support the new syntax that allows these to be things other
than array variables.  Some checking of shape compatibility is done.
At present only number of dims and total size is checkable.

Added new f77 option 'substring-bounds' which enforces F77 bounds
limits 1<=start<=end<=len for substrings, where len is the length of
the parent string.  F90 relaxes the rule to require only 1<=start and
start,end<=len.

Changed module file format for arrays: instead of writing the packed
(dims,size) bundle, it separately writes dims and size.  Also changed
handling of modules to allow array-valued parameters.  (Formerly array
and parameter were mutually exclusive attributes.)

In dtypes.c: fixed bug in last_component() to correctly handle
components that are array elements.  (Still needs to be fixed to
handle array sections when those are implemented.)

In symtab.c: fixed bug in call_subr that was not creating an inherited
local-scope copy of subroutine that exists in enclosing subprogram of
a module or internal subprogram.

In fortran.y: added syntax to support array sections.  The action code
for this is not done yet.  Repurposed the syntax for substring bounds
ranges so it can be shared with array sections.  This changed some
error messages, which are now given at integer_expr production and
in substring_size() in exprtype.c.