In symtab.h : Redesigned array_dim_t struct, removing size_unknown flag and changing size from unsigned to signed long, so that unknown size can be represented using size_UNKNOWN as for strings.
Renamed token flag ARRAY_ID_EXPR to ARRAY_EXPR since it identifies
array-valued items, which need not be identifiers now that
array-valued expressions are supported.

In exprtype.c : Added function check_array_conformance() which is used
in binary expressions and assignments involving array-valued
expressions. It currently checks for rank equality and total size
equality. When array shapes are implemented each dimension's size
equality will be checked.

In exprtype.c : Added function subarray_size() which calculates rank
and total size (number of elements) of a subarray.  Added function
array_section_size() which is called by subarray_size to figure out
the size of each dimension in a subarray.  If one of the range bounds
is empty, currently we cannot do the right thing to replace it by the
corresponding endpoint, because full array shape is not stored.

In exprtype.c : Extended function substring_size() to handle the
special case of substrings that refer to the same starting and ending
index variable, e.g. C(I:I) is length 1.  Also done for
array_section_size().  See ticket #94.

In fortran.y : Grammar now recognizes component references in which
the ultimate component is a substring array element.

In dtypes.c : Extended function last_component() to determine array
rank and size when base and/or components are arrays.  Also determines
substring size when last component is character type.  This needs to
be done in last_component() rather than by action code in the grammar
(as is done for plain arrays and strings) because the grammar does not
know whether components have array attribute or character type.