Updated README and PATCHES for release 4.alpha.2.
Updated configure.generic with substitutions for text-to-ps and
ps-to-pdf programs now needed by Makefile for listings target
introduced in previous version.

Added rule to test/Makefile.in to create test/Okay.dos directory if it
does not exist.  This directory is not in repository and so needs to
be created after a fresh checkout is set up.  It will be created when
the "pretar" target makes the DOS versions of regression output files.