In exprtype.c: fixed bug: ARRAY_EXPR not being propagated to result token in binexpr_type.  I think the same bug is in unexpr_type, not yet fixed.
In symtab.h and symtab.c: designed encoding of kind parameters,
default kinds and kinds specified by SELECTED_INT_KIND(RANGE) and
SELECTED_REAL_KIND([PRECISION],[RANGE]).  The encoding allows the type
of the parameter to be recovered (so we can warn if user defines a
parameter for a kind for one type and uses it for another type).  A
different kind number is produced for each precision and range
combination, except for small values that the F2010 standard specifies
should be supported by any default integer and real kinds, and for
very large values, which peg at upper limits.  Added handlers for the
kind intrinsic functions.  Note that until we support argument
call-by-name, there is no way to specify SELECTED_REAL_KIND(RANGE),
without PRECISION.

In intrins.c: added I_NONF90 flag to NULL intrinsic function, which is
in F95 but not F90.  We need to add a -f95 option to allow
differentiation between these standards.  (Likewise for F2010.)

In intrins.c: fixed bug in SIZE and LEN intrinsic function handlers,
that did not check for absence of an argument, which can happen on
illegal code and would cause seg fault.  (These inquiry functions are
invoked regardless of argument status.)

Began adding correct treatment of elemental functions.  These take
array properties from arguments rather than from symbol table
definition.  This is just at proof of principle stage right now.
II_ELEM flag needs to be added to elemental intrinsic functions in
table in intrins.c.  Only a few have it at the moment.  Also not yet
done is array conformance check (and promotion of scalars) on
arguments, similar to the way argument type equality is checked for
intrinsics.  But note user-defined functions can be elemental too.