Added rule to make tokdefs.h in case it does not exist, as is the case in a fresh checkout.  For some reason the "devel" target causes make to complain "No rule to make target `tokdefs.h', needed by `block_match.h'".  (The "prog" target does not.)  The rule has no dependency since we want to avoid re-making tokdefs.h unless the list of tokens has changed.  If the token definitions do change, it always means fortran.y has changed, so fortran.c will be remade and en passant tokdefs.h will be too.
Added a configure action to correct a change made in revision 32.  The
problem being addressed is that if blockmatch.h has a dependency on
make_blockmatch.pl, then it will be re-made after every configure
run.  This is bad for the distro version, since users may not have a
working Perl.  (Unlikely, but why look for trouble?)  However, if this
dependency is missing then if the developer makes a substantive change
to make_blockmatch.pl, make won't know it needs to re-make
blockmatch.h.  The correct solution is to have the --enable-devel
option of configure include the dependency in Makefile, but otherwise
configure leaves it out.

Changed dependency of "check" target in Makefile from "prog" (which
does not include development flags) to "default" (which will be either
"prog" or "devel" depending on --enable-devel).  Previously if "make
check" was run when ftnchek was out of date with respect to some but
not all sources, link could fail due to lack of -lm flag.