Parser now accepts FORALL statements and FORALL construct statements.
In symtab.c : Added function def_forall_index() which defines the
index variables of a FORALL statement.  Modification of FORALL index
variables is caught.  Added function process_forall_construct() which
warns if the FORALL index was unused in the construct or statement.
If it finds any variables erroneously in the local scope that are not
the index variables, it moves them to enclosing scope.