Implemented initial version of "lookahead kluge" that addresses the provides an explicit interface (actually just data type, size, and kind) of internal procedures at the point of invocation by host subprogram.  It does not yet get the correct type for functions with result variables, or find procedures invoked by peers within the same CONTAINS section.
Also fixed a pre-existing bug in regular lookahead routines used by
lexer.  They formerly assumed that end of statement would coincide
with a terminating null in the source line.  With the advent of
free-form continuations and inline comments, this assumption is no
longer correct.  The routines now use a test for index in line greater
than end_index (which is last significant character of line).