Added files internalfuns.f90 charfuns.f90 to regression test suite.
Tweaked warnings for truncation in assignment so that (a) "may not
give desired precision" is issued only for floating-point numeric
types (it was being issued for character assignments), and (b) it says
"truncated" regardless if default kind if size_trunc is true for
character type, since kind does not matter for truncation in that
case.  (One presumes the differing kind specs are OK, e.g. kind=1 and
default, on the given processor.)