Finished (I think) the lookahead kluge, apart from embellishments such as trying harder to get the correct kind parameter.  It now accepts attribute declarations of function result, e.g. POINTER :: F_RESULT.
Removed code in parse_prefix_keywd() that advanced past comma after
CHARACTER*len.  The standard allows an optional comma in this
deprecated form, e.g. CHARACTER*8 C or CHARACTER*8, C .  This is
allowed only in non-attribute-based type declarations, since in those
declarations the comma is needed to separate the type spec from
following attributes.  It is too much trouble to look ahead and see if
there is a double colon or not, and we expect internal functions will
be written in modern syntax, so little harm should come from
disallowing the optional comma here.