Addressed part of ticket #75, which is to properly deal with multiple USE statements referring to the same module.  This can happen explicitly, when a program unit and/or its internal procedures USE module A twice, but it can also happen implicitly, when a program unit USEs modules A and B, while B itself also USEs A.
Added new field Gsymtab* home_unit to local symbol table entry, to
hold a pointer to the program unit in which the variable is defined.
This only matters for module variables, but for simplicity it is set
for all variables.  Added a new field "home" in the module file format
for variables, which is the name of the home unit.  Then in project.c,
mod_var_in(), when a variable is read from the module file, if it was
seen before and the existing entry has the same home unit as the new
one, no new symbol table entry is created.  If it is new, the
home_unit field in the symtab entry is set to the module named in the
"home" field.

This change causes a difference in module file format.

Note that the same treatment needs to be applied to module procedures.