Added "home" field to "entry" and "external" lines in module and project files, to record the module where a procedure is defined. (For project files this field is set to the name of the procedure itself, but is ignored.)  This field is used to prevent creation of duplicate, masking local and global symbol table entries for the procedure when a module is USEd more than once in a program unit or its internal procedures.
In project.c: Fixed bug that can affect comparisons of topfilename to
top_filename which are done by comparing pointers rather than strcmp.

Fixed bug in find_prog_units() for the selection of which global
symbol table entries to use: it had been using the criterion for
project files, that the procedure should be defined in the same file.
For modules, the scan of the global symbol table is limited to
procedures referenced in the module, so all of the ones with
definitions should be included.  This selection test is done by a new
routine for_mod().

Changed the Lsymtab home_unit field from Gsymtab* to char* and
pointing to the name of the home unit.  At the time install_local() is
called for a module or external subprogram, it has not global symbol
table entry.  Note also that a module procedure defined in the module
being written out does not have a local symbol table entry any more at
the time the module file is written.