Fixed bug in fortran.y, in production ordinary_stmt where the statement is the first one, implying a main program.  The test of whether to move the %MAIN symtab entry out of scope was incorrect for the case where the statement is an INTERFACE that, unlike TYPE, does not put anything into the symbol table.
Fixed some architecture-dependent mismatches between format specs and
output items in debugging code.

Fixed bug in get_punctuation() to properly recognize OPERATOR (/) as
four tokens and not be fooled into lexing the (/ or /) as array
constructor delimiters.

Fixed bug in fortran.y, in pop_block() erroneously issuing an error
message when END INTERFACE closing an INTERFACE with generic-spec
lacks the generic-spec.  This is not an error for TYPE or INTERFACE
blocks.  I noticed that for TYPE blocks, the error was suppressed
because the block name was saved in curr_stmt_name all the way from
the opening TYPE statement to the closing END TYPE statement.  There
should be no named constructs inside a type definition block but this
is unsafe and obscure.  Also it was incorrectly ignoring the
symbolic_name on the END TYPE statement if present.  I added action
code at end_derived_type_stmt to set curr_stmt_name to NULL if not
present and to the generic-spec if present.  Then in pop_block(),
must_check_name is set to FALSE for END TYPE and END INTERFACE
statements so the error message will not be issued if the block name
is missing.

Fixed bug causing seg fault for code having a statement like PUBLIC ::
OPERATOR(+).  We have not yet implemented storing of defined operators
in the local symbol table.  The action code for this production calls
apply_attr() with the OPERATOR token as argument, which is not the
right thing to do.  This created a defective symbol table entry.  For
now, apply_attr() simply ignores any argument that is not an ID_EXPR.
Eventually we need to create proper symbol table entries for operators
and assignments as part of the semantic processing of INTERFACE blocks.