Fixed a bug when intrinsic function used by internal/module procedure when host declares or uses it.  The info.arglist gets cleared by inherit_local() but for intrinsics this cleared info.intrins_info. Solution is for inherit_local() to leave info alone if intrinsic flag is set.
Fixed bug causing a warning "Disassociating an allocated pointer" when
nullifying a field of a structured type variable that was allocated,
e.g.
    type(t_type) t
    allocate (t)
    nullify (t % left)
The warning is now suppressed if the arg token has DTYPE_COMPONENT
flag set.

Fixed bug causing a warning "Variables may be used
unassociated/unallocated:" about a variable that is passed to a
subroutine where it gets allocated.  We have to assume actual args may
get allocated or associated.  The treatment of allocation and
association for actual args now follows that of being set.