In advance.c: Added ALLOCATABLE to attributes that lookahead kluge will recognize in function result type declarations.  This attribute is illegal for function result variables in F95 but legal in F2003, and gfortran accepts it, so we may as well allow it.  This attribute is not included as a field in ProcInterface structure since the allocatability is inside the function body, not in the caller.
In project.c: Fixed bug in storage of parameter values in module
files.  It had been storing the source text, which is correct for
strings and OK for numeric literal constants, but generally the source
text may be an expression, and we need the actual value if the
parameter is an evaluated constant.  (The value is read in using %d or
%f.) This changes the behavior for real parameters since these are not
evaluated, and they are stored as 0.0 regardless of value.

In intrins.c: Fixed errors in the definition of SIZE intrinsic that
prevented it from allowing the DIM argument.

In loccheck.c: suppressed warning about function modifying an
argument, if the argument has INTENT(OUT).  The -pure flag, which turns
on this warning, is on by default, so INTENT(OUT) arguments may be
present legitimately.  (This fix will need to be adjusted when PURE
attributes of functions are recognized, since a function that is
declared as PURE is not allowed to have INTENT(OUT) arguments.  See
ticket #131.)

In fortran.y: Fixed bug, where parameters used as kind selectors did
not get their used_flag set, both in type declarations and in
selectors on literal constants.  The productions for kind_selector did
not, as had been foreseen, shift from using expr to using
int_constant_expr to provide appropriate checking.  Also, warnings
related to using kind parameters for the wrong type, or using concrete
kinds, were only being issued for the (KIND=value) form, not the
(value) form.  These are now unified.