Uncommented a line in test/array_rank1.f90 that ftnchek can now handle correctly.
In intake.c: Fixed bug that caused bogus warnings "Continuation in character
context must resume with '&'".  Caused by failure to initialize
contin_wo_amp flag when creating a srcLine instance.

In prlists.c: Fixed bug, failure to propagate array_dim of non-id
expressions passed as actual arguments.

In project.c: Changed the way real and complex parameters are stored
in module files, to prevent contents of module file from depending on
the value of the NO_FLOATING_POINT preprocessor macro (which is turned
off by the DEVELOPMENT macro (via DEBUG_FORLEX) but is on for the
distributed source).  Since the value of real parameters is unneeded,
a fixed value 0.0 is written.  (For complex, only the real part is
stored even if floating point is turned on.)  Also added explicit
support for LOGICAL parameters, although when these would be useful is
not obvious.  Fixed a bug in the reading of STRING parameters, which
came in when the previous bug fix shifted what is stored from the
source text to the stored parameter value (which lacks the string
delimiters).  Note that STRING parameters will not be correctly
recovered from the module file if (a) they exceed MAXNAME characters
in length or (b) they contain leading, trailing, or embedded blanks.
This may matter if a specification uses the length of a string
parameter.  See ticket #132.

In project.c: Minor improvement to last change, to avoid allocating
global storage twice for the same module variable name when it is not
renamed.

Added "const" qualifier to some pointer arguments, and cast some
fscanf() calls to (void), to suppress compiler warnings.

Removed the Linux Software Map file ftnchek-x.y.z.lsm from the
distribution.  According to ibiblio.org, new contributions are no
longer accepted, so ftnchek-4 will not be put there.

Began updating the README file for the next release, which will be
PETTT Deliverable 03.