Changed module file format to store array dim info for functions returning an array valued result.  In order for the information stored to be correct, changes were also necessary in def_array_dim() to copy the dim info of a result variable to its function, and to the ArgListHeader struct to add fields to hold the array dim info.  Then in process_lists() the info from the symbol table entry is copied into the arglist header, and alist_out() uses that for the module file. Note that both defn and call have their own sources of array dim information: the defn from the declarations in the function, and the call via whatever interface information the caller has.  Thus it is appropriate to put this information in the arglist header line rather than in the entry line of the module file (where elemental, pure, and recursive attributes go).
Also added storing of recursive attribute of procedures in module
file.

Also fixed a bug in debug_symtabs() causing incorrect operation,
including sometimes seg fault.  This bug actually dates back to 2001
when the -mkhtml flag was introduced.  In process_lists(), the
info.toklist field of a subprogram local symbol table entry gets
replaced by the newly created info.arglist, for use by make_html().
The thinking was that the toklist is no longer used, but the debug
output uses it.  Since it thinks the arglist is a toklist, it
misinterprets the struct pointed to.  The fix is simply to make the
replacement of toklist by arglist conditional on the html_documents
flag and, in the unlikely case that both -mkhtml and -local are used
at once, inform the user that this combination prohibits listing of
argument lists.