Added new commandline option -module=DIR which specifies a directory to which module files are written, and/or in which to search for module files for input.  For input, it now searches for module in the directory of the source file, then in DIR specified by -module option, then same search sequence as for include-files.
Removed all dependencies on #ifdef ALLOW_INCLUDE because the new
search scheme for modules uses find_include(), and the code would no
longer build if ALLOW_INCLUDE were to be turned off.  (The macro was
defined as 1 unconditionally for a long time now, but in theory one
could have edited ftnchek.h to change it to 0.  Now that is no longer
possible even in theory.)

Fixed a bug in read_module_file: fclose() was not being called
at the end, so file descriptor was left open.  This could cause
problems if a program uses many modules, making it run up against the
limit on the number of open file descriptors.