Began updating ftnchek.man to reflect the new version 4.  It is more or less complete through the OPTIONS section.  Major work remaining is the Limitations and Extensions section.
This commit was supposed to be mainly about the documentation, but
in the process of revising, some things were found that I felt should
be fixed rather than documented.  Details follow.

Removed all -f77 command-line options referring to Fortran 90 syntax
features, and added a single -f77=f90 option to cover them all.  There
are far too many of these features for it to be reasonable to define
options for all.  The only previous -f77 options remaining are for
vendor extensions that were not adopted in Fortran 90.  To make this
transition as simple as possible, the flag variables for the removed
options are all #define'd to f77_f90, so they are still distinguished
in the code that tests them, but they are no longer separately
controllable.  In some cases the option remains but has less scope
than before:

  -f77=format-edit-descr now only controls warnings about some edit
   descriptors that are non Fortran 90.

  -f77=long-names now only controls warnings about identifiers longer
   than 31 characters.

  -f77=typeless-constant now only controls the warnings
  about the non-f90 forms such as '123'Z whereas an f90 form like Z'123'
  is controlled by -f77=f90.  (It also now controls warnings about such
  constants appearing outside a DATA statement, which gfortran accepts
  by default.)

  -f77=initializer now controls only warnings about the nonstandard
  DATA-like form name /value/, not the F90 standard name=expr form.

  -f77=internal-list-io is now -f77=internal-namelist-io and only
  warns if I/O to internal file uses a namelist; warning about using
  list-directed io is controlled by -f77=f90.

  -f77=intrinsic now controls only warnings about vendor extension
   intrinsics, not F90 intrinsics.

I have some doubts about this new scheme, mainly because now except
for -f77=f90, the -f77 options cover the same things as the -f90
options.  But it seems to me that the other way to go, having the
above options control warnings about all non-F77 aspects, including
both F90 and non-F90 aspects, would make the system confusing for
users.

The above changes led to addition of new warnings related to I/O:
syntax error if format and namelist are both specified or if a
namelist I/O statement has an I/O list, and new
-f90=internal-namelist-io controlling warning about namelist I/O to
internal file.  This last is treated as a vendor extension rather than
a syntax error because gfortran by default accepts it.
In fortran.y: added syntax error message for use of internal file in
unformatted I/O.

Added new -f90=common-subprog-name option to perform the same check as
-f77=common-subprog-name, since it is still illegal for a common block
to have the same name as an external subprogram.

Withdrew top-level command-line switches -reference and -vcg, which
are options within -calltree, and -volatile which is an option of
-common .  There is no need for these to have their own top-level
switches.

Changed definition of the -identifier-chars option, moving underscore
out of it into the list of standard identifier characters.  Only
dollar sign remains by default, and now using -no-identifier will
leave recognition of underscore in place.  (It will still be warned
about if -f77=f90 is used.)

Fixed bug that goes back at least to version 3, failing to catch the
VMS Fortran usage OPEN(NAME=filename...).  Both f77 and f90 require
FILE instead of NAME.

Added files to test suite: open_with_name.f to catch use of OPEN
with NAME keyword, internal_list_io.f to catch errors involving I/O to
internal file, and boz_constants.f90 to test warnings about BOZ
aka typeless constants.

In test/Makefile.in, changed checkrc target actions because changes to
the -f77 options rendered the existing actions no longer useful.

Fixed bugs in add_html_refs.sh and make_html.sh, which were invoking
ftnchek via the user's PATH, rather than the one in the local
directory.  This causes problems when the options change.  In
add_html_refs.sh added deletion of the "Return to Main Contents" link
put in by man2html, which links to 

Added new option -common=name to produce warnings when corresponding
variables in different declarations of the same block differ in name.
This is expected to be useful when porting code to F90 and moving the
variables into modules.  At that point all subprograms need to be
using the same names in their code.