Minor changes resulting from review of gcc's "may be used uninitialized" warnings.  (These only appear when optimization such as -O2 is used.)  In some cases, code was changed to make the compiler stop warning (e.g. re-defining a variable that was previously defined inside an if with the same test).  In other cases comments were added to inform the reader that there is no error.  These are of the form /*LINTED*/ /*not uninit*/ which, according to the lint man page, suppresses lint warnings if placed on the line preceding the warning. Unfortunately gcc does not recognize these comments.  (The "not uninit" is in a separate comment to explain to the reader what type of warning the /*LINTED*/ is for.)
One actual bug turned up in func_ref_expr() in exprtype.c, in which
the defn struct for intrinsic functions was being accessed in a
statement that was not restricted to intrinsic functions.  This has
been fixed.  Another perhaps not a bug but it was hard to be sure:
using result_type in integer division truncation warning, which
seemed to belong better inside the else (result_type != E) block.

Also added an informational message to configure.in to help user know
what to do if man-to-html converter is not found.