Added ftnchek.pdf to DOCS macro and docs target in Makefile.in so it will be made and included in distribution tarball.  It is created from ftnchek.ps by the PStoPDF converter.
Made some more changes to ftnchek.man, mainly improving wording and
fixing issues with the HTML version.

Added ftnchek.html.patch to svn control.  This is somewhat of an
oddity for version control since it is not edited directly.  But it is
needed for the htmldocs target in Makefile.in to produce the right
thing after an update when ftnchek.man has been edited.  The way this
works is that the developer edits the primary documentation file
ftnchek.man, which is made into HTML by the MANtoHTML converter.  This
initial HTML output is placed in a file named ftnchek.html.orig.  This
file is then patched using ftnchek.html.patch to make minor fixes, and
the result is placed in ftnchek.html.  (An example of such fixes is
undoing bogus hyperlinks on text like SUBA(1) which man2html reads as
a man page reference and converts into a link to the CGI manpage for
the nonexistent program SUBA.)  Then make_html.sh is run on
ftnchek.html to produce the broken-out pages in the html/ directory.
Initially ftnchek.html.patch does not exist and ftnchek.html is a copy
of ftnchek.html.orig.  The developer edits ftnchek.html to taste, then
does "make ftnchek.html.patch" which saves the edits to be applied
again whenever ftnchek.man changes and htmldocs gets remade.  It would
not be right to put ftnchek.html into version control since it gets
overwritten every time htmldocs is remade.  The patch file, on the
other hand, is relatively stable (and also, of course, much smaller,
containing only the edits and a bit of surrounding context).  In sum,
the steps are:
1.  edit ftnchek.man for content
2.  run "make docs"
3.  view html in browser and look for issues
4.  edit ftnchek.html and run "make htmldocs" till html looks ok
5.  run "make ftnchek.html.patch" to save edits
Can now return to steps 1 & 2 without losing edits.